# swagger.yaml
basePath: /products
consumes:
    - application/json
definitions:
    Product:
        description: 'swagger: model Product'
        properties:
            description:
                description: The description for this product
                maxLength: 10000
                type: string
                x-go-name: Description
            id:
                description: The ID for this product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: The name for this product
                maxLength: 255
                type: string
                x-go-name: Name
            price:
                description: The price for the product
                format: double
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: The SKU for the product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - price
            - sku
        type: object
        x-go-package: go-microservices/product-api/data
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Return a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
    /products/{id}:
        put:
            description: Return a list of products
            operationId: updateProducts
            parameters:
                - description: id of the product to be updated
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    $ref: '#/responses/updatedSuccessfully'
            tags:
                - products
responses:
    productsResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
    updatedSuccessfully:
        description: ""
schemes:
    - http
swagger: "2.0"